+++ TowerOffense.Scenes/ExampleScene/ExampleScene.cs
@@ -1,16 +1,16 @@
 using Microsoft.Xna.Framework;
-using TowerOffense.Scenes.TestScene.Objects;
+using TowerOffense.Scenes.ExampleScene.Objects;
 using TowerOffense.Window;
 
-namespace TowerOffense.Scenes.TestScene {
+namespace TowerOffense.Scenes.ExampleScene {
     public class ExampleScene : Scene {
         public ExampleScene() : base() {
-            var window = new TestWindow(this, 200, 200);
+            var window = new ExampleWindow(this, 200, 200);
             window.ClearColor = Color.Black;
             window.Position = Point.Zero;
             AddObject(window);
 
-            var childWindow = new ChildTestWindow(this, 120, 120);
+            var childWindow = new ChildExampleWindow(this, 120, 120);
             childWindow.ClearColor = Color.Black;
             childWindow.Position = Point.Zero;
             AddObject(childWindow);
