+++ TowerOffense.Scenes/ExampleScene/Objects/ExampleWindow.cs
@@ -1,9 +1,9 @@
 using Microsoft.Xna.Framework;
 using TowerOffense.Window;
 
-namespace TowerOffense.Scenes.TestScene.Objects {
-    public class TestWindow : WindowObject {
-        public TestWindow(Scene scene, int width, int height) : base(scene, width, height) { }
+namespace TowerOffense.Scenes.ExampleScene.Objects {
+    public class ExampleWindow : WindowObject {
+        public ExampleWindow(Scene scene, int width, int height) : base(scene, width, height) { }
         public override void Render(GameTime gameTime) { }
         public override void Update(GameTime gameTime) { }
     }
